Refactor: Move ProviderEntity to provider module
Moved the `ProviderEntity` class from `com.currency.rateman.data.db.entity` to `com.currency.rateman.provider.data.entity`.

Updated all references to `ProviderEntity` to reflect the new location.
