+++ app/src/main/java/com/currency/rateman/data/db/entity/CurrencyRateEntity.kt
@@ -3,6 +3,7 @@ package com.currency.rateman.data.db.entity
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
+import com.currency.rateman.provider.data.entity.ProviderEntity
 
 @Entity(
     tableName = "currency_rates",
