+++ app/src/main/java/com/currency/rateman/provider/data/dao/ProviderDao.kt
@@ -5,7 +5,7 @@ import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
-import com.currency.rateman.data.db.entity.ProviderEntity
+import com.currency.rateman.provider.data.entity.ProviderEntity
 import kotlinx.coroutines.flow.Flow
 
 @Dao
