+++ app/src/main/java/com/currency/rateman/provider/data/repository/ProviderRepositoryImpl.kt
@@ -13,7 +13,7 @@ import com.currency.rateman.api.scraper.fetchRoyalExchangeRates
 import com.currency.rateman.api.scraper.fetchTopExchangeRates
 import com.currency.rateman.data.db.dao.CurrencyRateDao
 import com.currency.rateman.data.db.entity.CurrencyRateEntity
-import com.currency.rateman.data.db.entity.ProviderEntity
+import com.currency.rateman.provider.data.entity.ProviderEntity
 import com.currency.rateman.data.model.CurrencyRate
 import com.currency.rateman.data.model.enums.CurrencyCode
 import com.currency.rateman.data.model.enums.ProviderType
